+++ espresso/core/java/androidx/test/espresso/action/AdapterDataLoaderAction.java
@@ -20,7 +20,6 @@ import static androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.internal.util.Checks.checkNotNull;
 import static androidx.test.internal.util.Checks.checkState;
-import static kotlin.collections.CollectionsKt.mutableListOf;
 import static org.hamcrest.Matchers.allOf;
 
 import android.view.View;
@@ -32,6 +31,7 @@ import androidx.test.espresso.UiController;
 import androidx.test.espresso.ViewAction;
 import androidx.test.espresso.util.EspressoOptional;
 import androidx.test.espresso.util.HumanReadables;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import org.hamcrest.Matcher;
@@ -85,7 +85,7 @@ public final class AdapterDataLoaderAction implements ViewAction {
   @Override
   public void perform(UiController uiController, View view) {
     AdapterView<? extends Adapter> adapterView = (AdapterView<? extends Adapter>) view;
-    List<AdapterViewProtocol.AdaptedData> matchedDataItems = mutableListOf();
+    List<AdapterViewProtocol.AdaptedData> matchedDataItems = new ArrayList<>();
 
     for (AdapterViewProtocol.AdaptedData data :
         adapterViewProtocol.getDataInAdapterView(adapterView)) {