+++ espresso/core/javatests/androidx/test/espresso/remote/TestTypes.java
@@ -16,8 +16,6 @@
  */
 package androidx.test.espresso.remote;
 
-import static kotlin.collections.CollectionsKt.listOf;
-
 import android.net.Uri;
 import androidx.test.espresso.remote.annotation.RemoteMsgConstructor;
 import androidx.test.espresso.remote.annotation.RemoteMsgField;
@@ -28,6 +26,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -49,7 +48,7 @@ public final class TestTypes {
   static final TestType A_TEST_TYPE3 = new TestType(TEST_TYPE_STRING3);
   static final TestType A_TEST_TYPE4 = new TestType(TEST_TYPE_STRING4);
   static final Iterable<TestType> ANY_TYPE_ITERABLE =
-      listOf(A_TEST_TYPE1, A_TEST_TYPE2, A_TEST_TYPE3, A_TEST_TYPE4);
+      Arrays.asList(A_TEST_TYPE1, A_TEST_TYPE2, A_TEST_TYPE3, A_TEST_TYPE4);
 
   private TestTypes() {
     // no instance