+++ espresso/remote/java/androidx/test/espresso/remote/RemoteMessageSerializer.java
@@ -33,7 +33,6 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
-import kotlin.collections.CollectionsKt;
 
 /**
  * Serializes an arbitrary object into its proto message representation.
@@ -98,16 +97,13 @@ final class RemoteMessageSerializer implements EspressoRemoteMessage.To<MessageL
   private MessageLite toProtoInternal() {
     List<Field> targetFields = null;
     try {
+      // Transform fieldDescriptorList into a new list which contains field names
+      List<String> fieldNames = new ArrayList<>();
+      for (FieldDescriptor fieldDescriptor : fieldDescriptorList) {
+        fieldNames.add(fieldDescriptor.fieldName);
+      }
       // Filter a class declared fields based on field descriptor names
-      targetFields =
-          getFilteredFieldList(
-              instance.getClass(),
-              CollectionsKt.map(
-                  fieldDescriptorList,
-                  fieldDescriptor -> {
-                    // Transform fieldDescriptorList into a new list which contains field names
-                    return fieldDescriptor.fieldName;
-                  }));
+      targetFields = getFilteredFieldList(instance.getClass(), fieldNames);
       return createProtoFromTargetFields(targetFields, instance);
     } catch (Exception e) {
       if ((e instanceof RemoteProtocolException)) {