+++ espresso/remote/javatests/androidx/test/espresso/remote/ProtoReflectorTest.java
@@ -16,7 +16,6 @@
  */
 package androidx.test.espresso.remote;
 
-import static kotlin.collections.CollectionsKt.listOf;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -26,6 +25,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.SmallTest;
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
+import java.util.Arrays;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -58,7 +58,7 @@ public class ProtoReflectorTest {
     Any any3 = buildAny(ANY_TYPE_URL_THREE, ANY_TYPE_VALUE_THREE);
 
     Any anyValue1 = any1;
-    Iterable anyList1 = listOf(any2, any3);
+    Iterable anyList1 = Arrays.asList(any2, any3);
     ByteString value1 = TypeProtoConverters.typeToByteString(STRING_VALUE_ONE);
 
     ProtoReflector protoReflector =