+++ espresso/web/java/androidx/test/espresso/web/model/Atoms.java
@@ -17,12 +17,12 @@
 package androidx.test.espresso.web.model;
 
 import static androidx.test.internal.util.Checks.checkNotNull;
+import static java.util.Collections.emptyList;
 
 import androidx.annotation.VisibleForTesting;
 import androidx.test.espresso.remote.annotation.RemoteMsgConstructor;
 import androidx.test.espresso.remote.annotation.RemoteMsgField;
 import java.util.List;
-import kotlin.collections.CollectionsKt;
 
 /** Utility class wrapping simple and more commonly used atoms. */
 public final class Atoms {
@@ -53,7 +53,7 @@ public final class Atoms {
 
   /** Creates an atom that will execute the provided script and return an evaluation object. */
   public static Atom<Evaluation> script(String script) {
-    return new ScriptWithArgsSimpleAtom(script, CollectionsKt.listOf());
+    return new ScriptWithArgsSimpleAtom(script, emptyList());
   }
 
   /** Returns the value of document.location.href. */