+++ espresso/web/javatests/androidx/test/espresso/web/model/ModelCodecTest.java
@@ -16,17 +16,18 @@
 
 package androidx.test.espresso.web.model;
 
-import static kotlin.collections.CollectionsKt.listOf;
-import static kotlin.collections.CollectionsKt.mutableListOf;
+import static java.util.Collections.singletonList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.SmallTest;
 import com.google.common.collect.Maps;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import kotlin.collections.MapsKt;
 import org.json.JSONObject;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -71,7 +72,7 @@ public class ModelCodecTest {
 
   @Test
   public void testEncodeDecodeEvaluation_withList() {
-    List<Object> results = mutableListOf();
+    List<Object> results = new ArrayList<>();
     results.add(1);
     results.add(2);
     results.add(3);
@@ -94,7 +95,7 @@ public class ModelCodecTest {
 
   @Test
   public void testEncodeDecodeEvaluation_withMapValue() {
-    Map<String, Object> mapResult = MapsKt.mutableMapOf();
+    Map<String, Object> mapResult = new LinkedHashMap<>();
     mapResult.put("foo", "bar");
     mapResult.put("baz", true);
 
@@ -138,7 +139,7 @@ public class ModelCodecTest {
     adhoc.put("int_min", Integer.MIN_VALUE);
     adhoc.put("double_min", Double.MIN_VALUE);
     adhoc.put("double_max", Double.MAX_VALUE);
-    adhoc.put("sudz", listOf("goodbye"));
+    adhoc.put("sudz", singletonList("goodbye"));
     assertEquals(adhoc, ModelCodec.decode(ModelCodec.encode(adhoc)));
   }
 
@@ -150,7 +151,7 @@ public class ModelCodecTest {
     array[2] = Double.MIN_VALUE;
     array[3] = "Hello World";
     array[4] = 2;
-    assertEquals(listOf(array), ModelCodec.decode(ModelCodec.encode(array)));
+    assertEquals(Arrays.asList(array), ModelCodec.decode(ModelCodec.encode(array)));
   }
 
   @Test
@@ -159,9 +160,9 @@ public class ModelCodecTest {
     Map<String, Object> adhoc = Maps.newHashMap();
     adhoc.put("foobar", 1234);
     adhoc.put("buzzz", Boolean.FALSE);
-    adhoc.put("sudz", listOf("goodbye"));
+    adhoc.put("sudz", singletonList("goodbye"));
     List<Object> blah =
-        listOf(
+        Arrays.asList(
             (Object) new GoodJSONAble("something"),
             new GoodJSONAble(null),
             new ElementReference("haha"),
@@ -169,7 +170,7 @@ public class ModelCodecTest {
             Boolean.TRUE,
             42,
             null,
-            listOf((Object) "a nested list", 12345),
+            Arrays.asList((Object) "a nested list", 12345),
             adhoc,
             new GoodJSONAble("otherthing"));
     assertEquals(blah, ModelCodec.decode(ModelCodec.encode(blah)));